+++ src/r3/src/sync/mutex.rs
@@ -62,8 +62,7 @@ pub struct GenericMutex<Cell, Mutex> {
 ///
 /// const fn configure_app<C>(cfg: &mut Cfg<C>) -> Objects
 /// where
-///     C: ~const traits::CfgBase<System = System> +
-///        ~const traits::CfgTask +
+///     C: ~const traits::CfgTask<System = System> +
 ///        ~const traits::CfgMutex,
 /// {
 ///     StaticTask::define()
@@ -272,9 +271,7 @@ where
     System: traits::KernelMutex + traits::KernelStatic,
 {
     /// Complete the definition of a mutex, returning a reference to the mutex.
-    // `CfgMutex` can't have `~const CfgBase` as a supertrait because of
-    // [ref:const_supertraits], hence we need to specify `~const CfgBase` here
-    pub const fn finish<C: ~const traits::CfgMutex<System = System> + ~const traits::CfgBase>(
+    pub const fn finish<C: ~const traits::CfgMutex<System = System>>(
         self,
         cfg: &mut Cfg<C>,
     ) -> StaticMutex<System, Source::Target>