+++ src/r3_test_suite/src/kernel_benchmarks/timer_start.rs
@@ -35,9 +35,7 @@ impl<System: SupportedSystem> AppInner<System> {
     /// Used by `use_benchmark_in_kernel_benchmark!`
     const fn new<C, B: Bencher<System, Self>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System>
-            + ~const traits::CfgTask
-            + ~const traits::CfgTimer,
+        C: ~const traits::CfgTask<System = System> + ~const traits::CfgTimer,
     {
         // `<[T; LEN]>::from_fn` is not `const fn` [ref:const_array_from_fn]
         let timers = {