+++ src/r3_test_suite/src/kernel_tests/event_group_misc.rs
@@ -16,9 +16,7 @@ pub struct App<System: SupportedSystem> {
 impl<System: SupportedSystem> App<System> {
     pub const fn new<C, D: Driver<Self, System = System>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System>
-            + ~const traits::CfgTask
-            + ~const traits::CfgEventGroup,
+        C: ~const traits::CfgTask<System = System> + ~const traits::CfgEventGroup,
     {
         StaticTask::define()
             .start(task_body::<System, D>)