+++ src/r3_test_suite/src/kernel_tests/mutex_protect_priority_by_ceiling.rs
@@ -23,9 +23,7 @@ pub struct App<System: SupportedSystem> {
 impl<System: SupportedSystem> App<System> {
     pub const fn new<C, D: Driver<Self>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System>
-            + ~const traits::CfgTask
-            + ~const traits::CfgMutex,
+        C: ~const traits::CfgTask<System = System> + ~const traits::CfgMutex,
     {
         let task0 = StaticTask::define()
             .start(task0_body::<System, D>)