+++ src/r3_test_suite/src/kernel_tests/sync_mutex_misc.rs
@@ -32,8 +32,7 @@ pub struct App<System: SupportedSystem> {
 impl<System: SupportedSystem> App<System> {
     pub const fn new<C, D: Driver<Self>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System>
-            + ~const traits::CfgTask
+        C: ~const traits::CfgTask<System = System>
             + ~const traits::CfgInterruptLine
             + ~const traits::CfgMutex,
     {