+++ src/r3_test_suite/src/kernel_tests/time_misc.rs
@@ -19,7 +19,7 @@ pub struct App<System: SupportedSystem> {
 impl<System: SupportedSystem> App<System> {
     pub const fn new<C, D: Driver<Self>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System> + ~const traits::CfgTask,
+        C: ~const traits::CfgTask<System = System>,
     {
         StartupHook::define()
             .start(startup_hook::<System, D>)