+++ src/r3_test_suite/src/kernel_tests/timer_misc.rs
@@ -42,9 +42,7 @@ pub struct App<System: SupportedSystem> {
 impl<System: SupportedSystem> App<System> {
     pub const fn new<C, D: Driver<Self, System = System>>(b: &mut Cfg<C>) -> Self
     where
-        C: ~const traits::CfgBase<System = System>
-            + ~const traits::CfgTask
-            + ~const traits::CfgTimer,
+        C: ~const traits::CfgTask<System = System> + ~const traits::CfgTimer,
     {
         let timer1 = StaticTimer::define()
             .active(true)